requirements-dev.txt
-r requirements-test.txt

# Development dependencies compatible with Python 3.13
pytest-env==1.0.0
responses==0.25.0
pytest-watch==4.2.0
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
types-requests==2.32.0.20250602
types-waitress==3.0.1.20241117
